Added subtile constructor for btile. Rendering pipeline almost done.
Added lots of TODOs to handle a pile of edge cases.
Need to add to the bsm-space object the notion of placing a view at a location
in the buffer-space. The bsm-views current rendering is just a test, I need
to add the ability to increase the size of the views and let the view
control how many characters it wants to devote to a crate. I gotta think a
bit about a good heuristic to minimize the width and maximize the crates
rendered in the view, but keep comprehension high.