Major changes to rendering.
Each crate and view (and soon space) now has its own btile. This btile is
laid out by the parent. So a view figures out the locations and size for
a crate's btile then sets the crate's btile to be where it is supposed to be
and how big. A lot of small changes happened to effect this change.

Added much more rendering capability.

Added M-<left,right,up,down> keys which resize the view.
Added C-<left,right,up,down> which zoom in and out in the crate's dimensions.